requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
alembic
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
